chore: drop serde_utils and refresh callgraph
