+++ .gitignore
@@ -14,3 +14,4 @@ errors.log
 .coverage
 htmlcov/
 coverage.xml
+docs/callgraph.dot
