+++ src/moderation.py
@@ -6,6 +6,7 @@ import json
 import ast
 
 from log_utils import get_logger
+from config_utils import load_config
 from post_io import (
     read_post,
     raw_post_path,
@@ -18,9 +19,14 @@ from lot_io import read_lots, get_seller, get_timestamp
 
 log = get_logger().bind(module=__name__)
 
+try:
+    cfg = load_config()
+except SystemExit:
+    cfg = object()
+
 # Phrases that indicate spam or irrelevant posts.  The check is case insensitive
 # so new variations are still caught.
-BANNED_SUBSTRINGS = [
+DEFAULT_BANNED_SUBSTRINGS = [
     "прошу подпишитесь на канал @flats_in_georgia чтобы я пропускал ваши сообщения в этот чат!",
     "нарушил допустимую частоту публикации обьявлений и не сможет писать до",
     "вы используете запрещенное слово",
@@ -34,12 +40,15 @@ BANNED_SUBSTRINGS = [
     "mdma",
     "cтоимость рекламного пакета в нашей группе по тематике",
     "желающие снять квартиру в лучших, надежных и эффективных группах могут подписаться"
+
 ]
+
+BANNED_SUBSTRINGS = getattr(cfg, "BANNED_SUBSTRINGS", DEFAULT_BANNED_SUBSTRINGS)
 
 # Telegram usernames that only post housekeeping messages.  Their updates are
 # ignored entirely so we do not waste time downloading captcha images or other
 # irrelevant content.  All names must be lower case for easy comparison.
-BLACKLISTED_USERS = [
+DEFAULT_BLACKLISTED_USERS = [
     "m_s_help_bot",
     "chatkeeperbot",
     "dosvidulibot",
@@ -56,6 +65,10 @@ BLACKLISTED_USERS = [
     'verifuma_bot',
     'razvitiekanala_bot',
     'aboniment_admin1'
+]
+
+BLACKLISTED_USERS = [
+    u.lower() for u in getattr(cfg, "BLACKLISTED_USERS", DEFAULT_BLACKLISTED_USERS)
 ]
 
 LOTS_DIR = Path("data/lots")
